+++ bin/sysadmin/scriptshell/installer-sfynx-nginx.sh
@@ -3,6 +3,8 @@
 $INSTALL_USERWWW="/var/www"
 PLATEFORM_PROJET_GIT="https://github.com/pigroupe/cmf-sfynx.git"
 PLATEFORM_PROJET_NAME="sfynxproject"
+DATABASE_NAME="symfony"
+DATABASE_NAME_TEST="symfony_test"
 
 if [ $# -eq 0 ]; then # s'il n'y a pas de paramètres
     read INSTALL_USERWWW # on saisis la valeur
@@ -33,30 +35,41 @@ mkdir -p app/logs
 mkdir -p app/cachesfynx/loginfailure
 mkdir -p web/uploads/media
 mkdir -p web/yui
-rm app/config/parameters.yml
-cp app/config/parameters.yml.dist app/config/parameters.yml
+if [ ! -f app/config/parameters.yml ]; then
+    cp app/config/parameters.yml.dist app/config/parameters.yml
+fi
 
 # we add env var
 if ! grep -q "SYMFONY__DATABASE__NAME__ENV" ~/.profile; then
 cat <<EOT >> ~/.profile
 
 # env vars for SFYNX platform
-export SYMFONY__DATABASE__NAME__ENV=sfynx$PLATEFORM_PROJET_NAME_dev;
+export SYMFONY__DATABASE__NAME__ENV=$DATABASE_NAME;
 export SYMFONY__DATABASE__USER__ENV=root;
 export SYMFONY__DATABASE__PASSWORD__ENV=pacman;
-export SYMFONY__TEST__DATABASE__NAME__ENV=sfynx$PLATEFORM_PROJET_NAME_test;
+export SYMFONY__TEST__DATABASE__NAME__ENV=$DATABASE_NAME_TEST;
 export SYMFONY__TEST__DATABASE__USER__ENV=root;
 export SYMFONY__TEST__DATABASE__PASSWORD__ENV=pacman;
 EOT
 source ~/.profile
 fi
 
-# we create the virtualhiost of sfynx for nginx
-mkdir -p /tmp
-cat <<EOT >/tmp/$PLATEFORM_PROJET_NAME
+# On déclarer le socket Unix de PHP-FPM pour que Nginx puisse passer les requêtes PHP via fast_cgi
+if [ ! -f /etc/nginx/conf.d/php5-fpm.conf ];
+then
+sh -c "cat > /etc/nginx/conf.d/php5-fpm.conf" <<EOT
 upstream php5-fpm-sock {  
     server unix:/var/run/php5-fpm.sock;  
 }
+EOT
+fi
+
+# we create the virtualhiost of sfynx for nginx
+mkdir -p /tmp
+cat <<EOT >/tmp/$PLATEFORM_PROJET_NAME
+#upstream php5-fpm-sock {  
+#    server unix:/var/run/php5-fpm.sock;  
+#}
 
 server {
     set \$website_root "$INSTALL_USERWWW/$PLATEFORM_PROJET_NAME/web";
@@ -137,10 +150,10 @@ server {
         fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name;
         fastcgi_param  HTTPS off;
         # fastcgi_param PHP_VALUE "auto_prepend_file=$INSTALL_USERWWW/xhprof/external/header.php \n auto_append_file=$INSTALL_USERWWW/xhprof/external/footer.php";
-        fastcgi_param SYMFONY__DATABASE__NAME__ENV symfony_dev;
+        fastcgi_param SYMFONY__DATABASE__NAME__ENV $DATABASE_NAME;
         fastcgi_param SYMFONY__DATABASE__USER__ENV root;
         fastcgi_param SYMFONY__DATABASE__PASSWORD__ENV pacman;
-        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV symfony_test;
+        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV $DATABASE_NAME_TEST;
         fastcgi_param SYMFONY__TEST__DATABASE__USER__ENV root;
         fastcgi_param SYMFONY__TEST__DATABASE__PASSWORD__ENV pacman;
     }
@@ -261,10 +274,10 @@ server {
         fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name;
         fastcgi_param  HTTPS off;
         # fastcgi_param PHP_VALUE "auto_prepend_file=$INSTALL_USERWWW/xhprof/external/header.php \n auto_append_file=$INSTALL_USERWWW/xhprof/external/footer.php";
-        fastcgi_param SYMFONY__DATABASE__NAME__ENV symfony_dev;
+        fastcgi_param SYMFONY__DATABASE__NAME__ENV $DATABASE_NAME;
         fastcgi_param SYMFONY__DATABASE__USER__ENV root;
         fastcgi_param SYMFONY__DATABASE__PASSWORD__ENV pacman;
-        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV symfony_test;
+        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV $DATABASE_NAME_TEST;
         fastcgi_param SYMFONY__TEST__DATABASE__USER__ENV root;
         fastcgi_param SYMFONY__TEST__DATABASE__PASSWORD__ENV pacman;
     }
@@ -385,10 +398,10 @@ server {
         fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name;
         fastcgi_param  HTTPS off;
         # fastcgi_param PHP_VALUE "auto_prepend_file=$INSTALL_USERWWW/xhprof/external/header.php \n auto_append_file=$INSTALL_USERWWW/xhprof/external/footer.php";
-        fastcgi_param SYMFONY__DATABASE__NAME__ENV symfony_dev;
+        fastcgi_param SYMFONY__DATABASE__NAME__ENV $DATABASE_NAME;
         fastcgi_param SYMFONY__DATABASE__USER__ENV root;
         fastcgi_param SYMFONY__DATABASE__PASSWORD__ENV pacman;
-        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV symfony_test;
+        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV $DATABASE_NAME_TEST;
         fastcgi_param SYMFONY__TEST__DATABASE__USER__ENV root;
         fastcgi_param SYMFONY__TEST__DATABASE__PASSWORD__ENV pacman;
     }
@@ -430,10 +443,10 @@ server {
 
 }
 EOT
-mv /tmp/$PLATEFORM_PROJET_NAME /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME
+sudo mv /tmp/$PLATEFORM_PROJET_NAME /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME
 
 # we create the symbilic link
-ln -s /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME /etc/nginx/sites-enabled/$PLATEFORM_PROJET_NAME
+sudo ln -s /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME /etc/nginx/sites-enabled/$PLATEFORM_PROJET_NAME
 
 # we add host in the /etc/hosts file
 if ! grep -q "dev.$PLATEFORM_PROJET_NAME.local" /etc/hosts; then
@@ -453,6 +466,20 @@ if [ ! -f composer.phar ]; then
 fi
 php -d memory_limit=1024M composer.phar install --no-interaction
 php composer.phar dump-autoload --optimize
+
+#
+rm -rf app/cache/*
+rm -rf app/logs/*
+
+# Utiliser l'ACL sur un système qui supporte chmod +a
+HTTPDUSER=`ps aux | grep -E '[a]pache|[h]ttpd|[_]www|[w]ww-data|[n]ginx' | grep -v root | head -1 | cut -d\  -f1`
+sudo chmod +a "$HTTPDUSER allow delete,write,append,file_inherit,directory_inherit" app/cache app/logs
+sudo chmod +a "`whoami` allow delete,write,append,file_inherit,directory_inherit" app/cache app/logs
+
+# Utiliser l'ACL sur un système qui ne supporte pas chmod +a
+HTTPDUSER=`ps aux | grep -E '[a]pache|[h]ttpd|[_]www|[w]ww-data|[n]ginx' | grep -v root | head -1 | cut -d\  -f1`
+sudo setfacl -R -m u:"$HTTPDUSER":rwX -m u:`whoami`:rwX app/cache app/logs
+sudo setfacl -dR -m u:"$HTTPDUSER":rwX -m u:`whoami`:rwX app/cache app/logs
 
 # create database
 php app/console doctrine:database:create
