+++ provisioners/shell/plateform/installer-sfynx-nginx.sh
@@ -31,8 +31,9 @@ mkdir -p app/logs
 mkdir -p app/cachesfynx/loginfailure
 mkdir -p web/uploads/media
 mkdir -p web/yui
-rm app/config/parameters.yml
-cp app/config/parameters.yml.dist app/config/parameters.yml
+if [ ! -f app/config/parameters.yml ]; then
+    cp app/config/parameters.yml.dist app/config/parameters.yml
+fi
 
 # we add env var
 if ! grep -q "SYMFONY__DATABASE__NAME__ENV" ~/.profile; then
@@ -428,10 +429,10 @@ server {
 
 }
 EOT
-mv /tmp/$PLATEFORM_PROJET_NAME /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME
+sudo mv /tmp/$PLATEFORM_PROJET_NAME /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME
 
 # we create the symbilic link
-ln -s /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME /etc/nginx/sites-enabled/$PLATEFORM_PROJET_NAME
+sudo ln -s /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME /etc/nginx/sites-enabled/$PLATEFORM_PROJET_NAME
 
 # we add host in the /etc/hosts file
 if ! grep -q "dev.$PLATEFORM_PROJET_NAME.local" /etc/hosts; then
@@ -451,6 +452,20 @@ if [ ! -f composer.phar ]; then
 fi
 php -d memory_limit=1024M composer.phar install --no-interaction
 php composer.phar dump-autoload --optimize
+
+#
+rm -rf app/cache/*
+rm -rf app/logs/*
+
+# Utiliser l'ACL sur un système qui supporte chmod +a
+HTTPDUSER=`ps aux | grep -E '[a]pache|[h]ttpd|[_]www|[w]ww-data|[n]ginx' | grep -v root | head -1 | cut -d\  -f1`
+sudo chmod +a "$HTTPDUSER allow delete,write,append,file_inherit,directory_inherit" app/cache app/logs
+sudo chmod +a "`whoami` allow delete,write,append,file_inherit,directory_inherit" app/cache app/logs
+
+# Utiliser l'ACL sur un système qui ne supporte pas chmod +a
+HTTPDUSER=`ps aux | grep -E '[a]pache|[h]ttpd|[_]www|[w]ww-data|[n]ginx' | grep -v root | head -1 | cut -d\  -f1`
+sudo setfacl -R -m u:"$HTTPDUSER":rwX -m u:`whoami`:rwX app/cache app/logs
+sudo setfacl -dR -m u:"$HTTPDUSER":rwX -m u:`whoami`:rwX app/cache app/logs
 
 # create database
 php app/console doctrine:database:create
